refactor: split postgres platform store; persist trajectory agent_type
Postgres platform split — the 2,602-line platform.rs becomes 11 domain
modules mirroring temper-store-turso's store/ layout (largest 598 lines):
mod (shared helpers + re-exports), projection, trajectory, policies, specs,
evolution, wasm, published_artifacts, ots, secrets, blobs. Bodies moved
verbatim; the crate's public API is re-exported unchanged.

Trajectory agent_type write-through — TrajectoryEntry has carried
agent_type since the field was added, but neither store's INSERT wrote it,
so per-agent-type forensics were silently empty:
- turso: column added via the idempotent ALTER loop; INSERT now binds it
- postgres: column existed in the initial CREATE; added
  ADD COLUMN IF NOT EXISTS plus versioned migration 0004 for databases
  bootstrapped before the column entered the schema; INSERT now binds it
- temper-server: all three TrajectorySink impls pass
  entry.agent_type.as_deref()

Also documents the turso append timeout contract on EventStore::append:
a timeout means the write's fate is unknown; callers retry with the same
expected_sequence and the UNIQUE constraint converts a landed prior
attempt into ConcurrencyViolation rather than a duplicate.

postgres 38 + turso 48 + server 452 tests green; full workspace suite
green. DST review: PASS. Code review: PASS (migration idempotency and
column/bind ordering independently verified).
