+++ crates/temper-store-turso/src/schema.rs
@@ -289,6 +289,8 @@ pub const ALTER_TRAJECTORIES_ADD_REQUEST_BODY: &str =
 pub const ALTER_TRAJECTORIES_ADD_INTENT: &str = "ALTER TABLE trajectories ADD COLUMN intent TEXT";
 pub const ALTER_TRAJECTORIES_ADD_MATCHED_POLICY_IDS: &str =
     "ALTER TABLE trajectories ADD COLUMN matched_policy_ids TEXT";
+pub const ALTER_TRAJECTORIES_ADD_AGENT_TYPE: &str =
+    "ALTER TABLE trajectories ADD COLUMN agent_type TEXT";
 
 /// Index on agent_id for agent-scoped trajectory queries.
 pub const CREATE_TRAJECTORIES_AGENT_INDEX: &str = "\
